add refactor multi wallet, remove mainnet wallet from celo
