+++ src/server/blockchain/CeloAdminWallet.js
@@ -1,5 +1,7 @@
-import { Wallet } from './AdminWalletV2'
+import { Web3Wallet } from './Web3Wallet'
 import conf from '../server.config'
 
-const walletConfig = { ...conf, ethereum: conf.celo, network: conf.env + '-celo' }
-export default new Wallet(walletConfig)
+const { celo, env } = conf
+const network = `${env}-celo`
+
+export default new Web3Wallet('CeloAdminWallet', conf, celo, network)
