+++ src/server/blockchain/Web3Wallet.js
@@ -3,39 +3,61 @@ import Crypto from 'crypto'
 import Web3 from 'web3'
 import HDKey from 'hdkey'
 import bip39 from 'bip39-light'
-import { get } from 'lodash'
+import get from 'lodash/get'
+import assign from 'lodash/assign'
 import * as web3Utils from 'web3-utils'
 import IdentityABI from '@gooddollar/goodcontracts/build/contracts/Identity.min.json'
 import GoodDollarABI from '@gooddollar/goodcontracts/build/contracts/GoodDollar.min.json'
-import UBIABI from '@gooddollar/goodcontracts/stakingModel/build/contracts/UBIScheme.min.json'
+import UBIABI from '@gooddollar/goodprotocol/artifacts/contracts/ubi/UBIScheme.sol/UBIScheme.json'
 import ProxyContractABI from '@gooddollar/goodcontracts/build/contracts/AdminWallet.min.json'
-import ContractsAddress from '@gooddollar/goodcontracts/releases/deployment.json'
-import ModelContractsAddress from '@gooddollar/goodcontracts/stakingModel/releases/deployment.json'
-import UpgradablesContractsAddress from '@gooddollar/goodcontracts/upgradables/releases/deployment.json'
+import ContractsAddress from '@gooddollar/goodprotocol/releases/deployment.json'
 import FaucetABI from '@gooddollar/goodcontracts/upgradables/build/contracts/FuseFaucet.min.json'
 
 import conf from '../server.config'
 import logger from '../../imports/logger'
 import { isNonceError, isFundsError } from '../utils/eth'
-import { requestTimeout } from '../utils/async'
+import { withTimeout } from '../utils/async'
 import { type TransactionReceipt } from './blockchain-types'
 
 import { getManager } from '../utils/tx-manager'
 import { sendSlackAlert } from '../../imports/slack'
 
-const log = logger.child({ from: 'AdminWallet' })
-
-const FUSE_TX_TIMEOUT = 15000 //should be confirmed after max 3 blocks (15sec)
 const defaultGas = 200000
-const defaultGasPrice = web3Utils.toWei('1', 'gwei')
+const FUSE_TX_TIMEOUT = 25000 // should be confirmed after max 5 blocks (25sec)
+const { estimateGasPrice } = conf
 const adminMinBalance = conf.adminMinBalance
+const defaultGasPrice = web3Utils.toWei(String(conf.defaultGasPrice), 'gwei')
 const defaultRopstenGasPrice = web3Utils.toWei('5', 'gwei')
+const defaultCeloGasPrice = (0.2 * 1e9).toFixed(0)
+
+export const web3Default = {
+  defaultBlock: 'latest',
+  defaultGasPrice,
+  transactionBlockTimeout: 5,
+  transactionConfirmationBlocks: 1,
+  transactionPollingTimeout: 30
+}
+
+export const getAuthHeader = rpc => {
+  const url = new URL(rpc)
+
+  if (!url.password) {
+    return []
+  }
+
+  return [
+    {
+      name: 'Authorization',
+      value: `Basic ${Buffer.from(`${url.username}:${url.password}`).toString('base64')}`
+    }
+  ]
+}
 
 /**
  * Exported as AdminWallet
  * Interface with blockchain contracts via web3 using HDWalletProvider
  */
-export class Wallet {
+export class Web3Wallet {
   web3: Web3
 
   mainnetWeb3: Web3
@@ -64,152 +86,143 @@ export class Wallet {
 
   nonce: number
 
-  constructor(mnemonic: string) {
-    this.mnemonic = mnemonic
+  mainnetAddresses = []
+
+  constructor(name, conf, ethereum = null, network = null) {
     this.addresses = []
     this.filledAddresses = []
-    this.mainnetAddresses = []
     this.wallets = {}
-    this.numberOfAdminWalletAccounts = conf.privateKey ? 1 : conf.numberOfAdminWalletAccounts
-    this.network = conf.network
-    this.networkIdMainnet = conf.ethereumMainnet.network_id
-    this.networkId = conf.ethereum.network_id
-    this.maxMainnetGasPrice = conf.maxGasPrice * 1000000000 //maxGasPrice is in gwei, convert to wei
+    this.conf = conf
+    this.mnemonic = this.conf.mnemonic
+    this.network = network || conf.network
+    this.ethereum = ethereum || conf.ethereum
+    this.networkId = ethereum.network_id
+    this.numberOfAdminWalletAccounts = this.conf.privateKey ? 1 : this.conf.numberOfAdminWalletAccounts
+    this.maxMainnetGasPrice = this.conf.maxGasPrice * 1000000000 // maxGasPrice is in gwei, convert to wei
+    this.log = logger.child({ from: `${name}/${this.networkId}` })
+
     this.ready = this.init()
   }
 
   getWeb3TransportProvider(): HttpProvider | WebSocketProvider {
     let provider
     let web3Provider
-    let transport = conf.ethereum.web3Transport
+    let transport = this.ethereum.web3Transport
+    const { log } = this
+
     switch (transport) {
       case 'WebSocket':
-        provider = conf.ethereum.websocketWeb3Provider
+        provider = this.ethereum.websocketWeb3Provider
         web3Provider = new Web3.providers.WebsocketProvider(provider)
         break
 
       case 'HttpProvider':
       default:
-        provider = conf.ethereum.httpWeb3Provider
+        provider = this.ethereum.httpWeb3Provider
+        const headers = getAuthHeader(provider)
         web3Provider = new Web3.providers.HttpProvider(provider, {
-          timeout: 15000
+          timeout: FUSE_TX_TIMEOUT,
+          headers
         })
         break
     }
-    log.debug({ conf, web3Provider, provider })
 
+    log.debug({ conf: this.conf, web3Provider, provider })
     return web3Provider
   }
 
   addWallet(account) {
     this.web3.eth.accounts.wallet.add(account)
     this.web3.eth.defaultAccount = account.address
-    this.mainnetWeb3.eth.accounts.wallet.add(account)
-    this.mainnetWeb3.eth.defaultAccount = account.address
     this.addresses.push(account.address)
     this.wallets[account.address] = account
   }
 
-  getMainnetWeb3TransportProvider(): HttpProvider | WebSocketProvider {
-    let provider
-    let web3Provider
-    let transport = conf.ethereumMainnet.web3Transport
-    switch (transport) {
-      case 'WebSocket':
-        provider = conf.ethereumMainnet.websocketWeb3Provider
-        web3Provider = new Web3.providers.WebsocketProvider(provider)
-        break
-
-      default:
-      case 'HttpProvider':
-        provider = conf.ethereumMainnet.httpWeb3Provider
-        web3Provider = new Web3.providers.HttpProvider(provider)
-        break
-    }
-    log.debug('mainnet', { web3Provider, provider })
-
-    return web3Provider
-  }
-
   async init() {
-    log.debug('Initializing wallet:', { conf: conf.ethereum, mainnet: conf.ethereumMainnet })
-    this.mainnetTxManager = getManager(conf.ethereumMainnet.network_id)
-    this.txManager = getManager(conf.ethereum.network_id)
-    this.web3 = new Web3(this.getWeb3TransportProvider(), null, {
-      defaultBlock: 'latest',
-      defaultGasPrice,
-      transactionBlockTimeout: 5,
-      transactionConfirmationBlocks: 1,
-      transactionPollingTimeout: 20
-    })
+    const { log } = this
 
-    this.mainnetWeb3 = new Web3(this.getMainnetWeb3TransportProvider(), null, {
-      defaultBlock: 'latest',
-      transactionBlockTimeout: 5,
-      transactionConfirmationBlocks: 1
-    })
+    log.debug('Initializing wallet:', { conf: this.ethereum })
 
-    if (conf.privateKey) {
-      let account = this.web3.eth.accounts.privateKeyToAccount(conf.privateKey)
-      this.web3.eth.accounts.wallet.add(account)
-      this.web3.eth.defaultAccount = account.address
-      this.mainnetWeb3.eth.accounts.wallet.add(account)
-      this.mainnetWeb3.eth.defaultAccount = account.address
+    this.txManager = getManager(this.ethereum.network_id)
+    this.web3 = new Web3(this.getWeb3TransportProvider(), null, web3Default)
+    assign(this.web3.eth, web3Default)
+
+    switch (this.networkId) {
+      default:
+      case 122:
+        this.gasPrice = defaultGasPrice
+      case 42220:
+        this.gasPrice = defaultCeloGasPrice
+    }
+
+    if (estimateGasPrice) {
+      await this.web3.eth
+        .getGasPrice()
+        .then(price => (this.gasPrice = price))
+        .catch(e => log.warn('failed to get gas price', e.message, e))
+    }
+
+    if (this.conf.privateKey) {
+      let account = this.web3.eth.accounts.privateKeyToAccount(this.conf.privateKey)
 
       this.address = account.address
       this.addWallet(account)
+
       log.info('Initialized by private key:', { address: account.address })
     } else if (this.mnemonic) {
-      let root = HDKey.fromMasterSeed(bip39.mnemonicToSeed(this.mnemonic, conf.adminWalletPassword))
+      let root = HDKey.fromMasterSeed(bip39.mnemonicToSeed(this.mnemonic, this.conf.adminWalletPassword))
+
       for (let i = 0; i < this.numberOfAdminWalletAccounts; i++) {
         const path = "m/44'/60'/0'/0/" + i
         let addrNode = root.derive(path)
         let account = this.web3.eth.accounts.privateKeyToAccount('0x' + addrNode._privateKey.toString('hex'))
+
         this.addWallet(account)
       }
+
       log.info('Initialized by mnemonic:', { address: this.addresses })
     }
 
     const adminWalletAddress = get(ContractsAddress, `${this.network}.AdminWallet`)
-    this.proxyContract = new this.web3.eth.Contract(ProxyContractABI.abi, adminWalletAddress)
-
     const adminWalletContractBalance = await this.web3.eth.getBalance(adminWalletAddress)
+
+    this.proxyContract = new this.web3.eth.Contract(ProxyContractABI.abi, adminWalletAddress, { from: this.address })
     log.info(`AdminWallet contract balance`, { adminWalletContractBalance, adminWalletAddress })
-    if (parseFloat(web3Utils.fromWei(adminWalletContractBalance, 'gwei')) < adminMinBalance * this.addresses.length) {
+
+    if (web3Utils.fromWei(adminWalletContractBalance, 'gwei') < adminMinBalance * this.addresses.length) {
       log.error('AdminWallet contract low funds')
       sendSlackAlert({ msg: 'AdminWallet contract low funds', adminWalletAddress, adminWalletContractBalance })
-      if (conf.env !== 'test' && conf.env !== 'development') process.exit(-1)
+      if (this.conf.env !== 'test' && this.conf.env !== 'development') process.exit(-1)
     }
 
     this.txManager.getTransactionCount = this.web3.eth.getTransactionCount
-    this.mainnetTxManager.getTransactionCount = this.mainnetWeb3.eth.getTransactionCount
 
     await this.txManager.createListIfNotExists(this.addresses)
 
-    if (conf.env !== 'production') await this.mainnetTxManager.createListIfNotExists(this.mainnetAddresses)
-
     log.info('Initialized wallet queue manager')
-    if (conf.topAdminsOnStartup) {
-      await this.topAdmins(0, conf.numberOfAdminWalletAccounts).catch(e => {
+
+    if (this.conf.topAdminsOnStartup) {
+      await this.topAdmins(0, this.conf.numberOfAdminWalletAccounts).catch(e => {
         log.warn('Top admins failed', { e, errMessage: e.message })
       })
     }
 
     const ps = this.addresses.map(async addr => {
       const balance = await this.web3.eth.getBalance(addr)
-
       const isAdminWallet = await this.isVerifiedAdmin(addr)
+
       if (isAdminWallet && parseFloat(web3Utils.fromWei(balance, 'gwei')) > adminMinBalance) {
         log.info(`admin wallet ${addr} balance ${balance}`)
         this.filledAddresses.push(addr)
-      } else log.warn('Failed adding admin wallet', { addr, balance, isAdminWallet, adminMinBalance })
+      }
 
-      if (conf.env !== 'production') {
+      if (this.conf.env !== 'production') {
         const mainnetBalance = await this.mainnetWeb3.eth.getBalance(addr)
+
         if (parseFloat(web3Utils.fromWei(mainnetBalance, 'gwei')) > adminMinBalance * 100) {
           log.info(`admin wallet ${addr} mainnet balance ${mainnetBalance}`)
           this.mainnetAddresses.push(addr)
-        } else log.warn('Failed adding mainnet admin wallet', { addr, mainnetBalance, adminMinBalance })
+        }
       }
     })
 
@@ -218,19 +231,15 @@ export class Wallet {
 
     if (this.filledAddresses.length === 0) {
       log.error('no admin wallet with funds')
+
       sendSlackAlert({
         msg: 'critical: no fuse admin wallet with funds'
       })
-      if (conf.env !== 'test' && conf.env !== 'development') process.exit(-1)
-    }
 
-    // if (this.mainnetAddresses.length === 0) {
-    //   sendSlackAlert({
-    //     msg: 'critical: no mainnet admin wallet with funds'
-    //   })
-    //   log.error('no admin wallet with funds for mainnet')
-    //   if (conf.env !== 'test') process.exit(-1)
-    // }
+      if (this.conf.env !== 'test' && this.conf.env !== 'development') {
+        process.exit(-1)
+      }
+    }
 
     this.address = this.filledAddresses[0]
 
@@ -245,13 +254,14 @@ export class Wallet {
       get(ContractsAddress, `${this.network}.GoodDollar`),
       { from: this.address }
     )
-    this.UBIContract = new this.web3.eth.Contract(UBIABI.abi, get(ModelContractsAddress, `${this.network}.UBIScheme`), {
+
+    this.UBIContract = new this.web3.eth.Contract(UBIABI.abi, get(ContractsAddress, `${this.network}.UBIScheme`), {
       from: this.address
     })
 
     this.faucetContract = new this.web3.eth.Contract(
       FaucetABI.abi,
-      get(UpgradablesContractsAddress, `${this.network}.FuseFaucet`),
+      get(ContractsAddress, `${this.network}.FuseFaucet`),
       {
         from: this.address
       }
@@ -262,8 +272,10 @@ export class Wallet {
         .balanceOf(this.address)
         .call()
         .then(parseInt)
+
       let nativebalance = await this.web3.eth.getBalance(this.address)
       this.nonce = parseInt(await this.web3.eth.getTransactionCount(this.address))
+
       log.debug('AdminWallet Ready:', {
         activeWallets: this.filledAddresses.length,
         activeMainnetWallets: this.mainnetAddresses.length,
@@ -278,8 +290,11 @@ export class Wallet {
     } catch (e) {
       log.error('Error initializing wallet', e.message, e)
 
-      if (conf.env !== 'test' && conf.env !== 'development') process.exit(-1)
+      if (this.conf.env !== 'test' && this.conf.env !== 'development') {
+        process.exit(-1)
+      }
     }
+
     return true
   }
 
@@ -289,14 +304,18 @@ export class Wallet {
    * @returns {Promise<String>}
    */
   async topAdmins(numAdmins: number): Promise<any> {
+    const { log } = this
+
     try {
       const { nonce, release, fail, address } = await this.txManager.lock(this.addresses[0])
+
       try {
         for (let i = 0; i < numAdmins; i += 50) {
           log.debug('topAdmins sending tx', { address, nonce, adminIdx: i })
           await this.proxyContract.methods.topAdmins(i, i + 50).send({ gas: '500000', from: address, nonce })
           log.debug('topAdmins success', { adminIdx: i })
         }
+
         release()
       } catch (e) {
         fail()
@@ -308,33 +327,21 @@ export class Wallet {
   }
 
   /**
-   * charge bonuses for user via `bonus` contract
-   * @param {string} address
-   * @param {string} amountInWei
-   * @param {object} event callbacks
-   * @returns {Promise<String>}
-   */
-  async redeemBonuses(address: string, amountInWei: string, { onReceipt, onTransactionHash, onError }): Promise<any> {
-    return this.sendTransaction(this.proxyContract.methods.awardUser(address, amountInWei), {
-      onTransactionHash,
-      onReceipt,
-      onError
-    })
-  }
-
-  /**
    * whitelist an user in the `Identity` contract
    * @param {string} address
    * @param {string} did
    * @returns {Promise<TransactionReceipt>}
    */
   async whitelistUser(address: string, did: string): Promise<TransactionReceipt | boolean> {
+    const { log } = this
     const isVerified = await this.isVerified(address)
 
     if (isVerified) {
       return { status: true }
     }
+
     let txHash
+
     try {
       const lastAuth = await this.identityContract.methods
         .lastAuthenticated(address)
@@ -368,6 +375,8 @@ export class Wallet {
   }
 
   async authenticateUser(address: string): Promise<TransactionReceipt> {
+    const { log } = this
+
     try {
       let encodedCall = this.web3.eth.abi.encodeFunctionCall(
         {
@@ -382,8 +391,10 @@ export class Wallet {
         },
         [address]
       )
+
       const transaction = await this.proxyContract.methods.genericCall(this.identityContract._address, encodedCall, 0)
       const tx = await this.sendTransaction(transaction, {}, { gas: 500000 })
+
       log.info('authenticated user', { address, tx })
       return tx
     } catch (exception) {
@@ -395,14 +406,18 @@ export class Wallet {
   }
 
   async getAuthenticationPeriod(): Promise<number> {
+    const { log } = this
+
     try {
       const result = await this.identityContract.methods
         .authenticationPeriod()
         .call()
         .then(parseInt)
+
       return result
     } catch (exception) {
       const { message } = exception
+
       log.warn('Error getAuthenticationPeriod', message, exception)
       throw exception
     }
@@ -414,6 +429,8 @@ export class Wallet {
    * @returns {Promise<TransactionReceipt>}
    */
   async blacklistUser(address: string): Promise<TransactionReceipt> {
+    const { log } = this
+
     const tx: TransactionReceipt = await this.sendTransaction(this.proxyContract.methods.blacklist(address)).catch(
       e => {
         log.error('Error blackListUser', e.message, e, { address })
@@ -430,6 +447,8 @@ export class Wallet {
    * @returns {Promise<TransactionReceipt>}
    */
   async removeWhitelisted(address: string): Promise<TransactionReceipt> {
+    const { log } = this
+
     const tx: TransactionReceipt = await this.sendTransaction(
       this.proxyContract.methods.removeWhitelist(address)
     ).catch(e => {
@@ -446,6 +465,8 @@ export class Wallet {
    * @returns {Promise<boolean>}
    */
   async isVerified(address: string): Promise<boolean> {
+    const { log } = this
+
     const tx: boolean = await this.identityContract.methods
       .isWhitelisted(address)
       .call()
@@ -453,15 +474,31 @@ export class Wallet {
         log.error('Error isVerified', e.message, e)
         throw e
       })
+
     return tx
   }
 
+  async getDID(address: string): Promise<string> {
+    const { log } = this
+
+    const tx: boolean = await this.identityContract.methods
+      .addrToDID(address)
+      .call()
+      .catch(e => {
+        log.error('Error getDID', e.message, e)
+        throw e
+      })
+
+    return tx
+  }
   /**
    *
    * @param {string} address
    * @returns {Promise<boolean>}
    */
   async isVerifiedAdmin(address: string): Promise<boolean> {
+    const { log } = this
+
     const tx: boolean = await this.proxyContract.methods
       .isAdmin(address)
       .call()
@@ -469,6 +506,7 @@ export class Wallet {
         log.error('Error isAdmin', e.message, e)
         throw e
       })
+
     return tx
   }
 
@@ -477,20 +515,28 @@ export class Wallet {
    * @param {string} address
    * @returns {PromiEvent<TransactionReceipt>}
    */
-  async topWallet(address: string, logger = log): PromiEvent<TransactionReceipt> {
-    let userBalance = await this.web3.eth.getBalance(address)
-    let maxTopWei = parseInt(web3Utils.toWei('1000000', 'gwei'))
-    let toTop = maxTopWei - userBalance
-    logger.debug('TopWallet:', { address, userBalance, toTop })
-    if (toTop <= 0 || toTop / maxTopWei < 0.75) {
-      logger.debug("User doesn't need topping", { address })
-      return { status: 1 }
+  async topWallet(address: string, customLogger = null): PromiEvent<TransactionReceipt> {
+    const logger = customLogger || this.log
+    const faucetRes = await this.topWalletFaucet(address, logger).catch(_ => false)
+
+    if (faucetRes) {
+      return faucetRes
     }
 
-    const faucetRes = await this.topWalletFaucet(address, logger).catch(_ => false)
-    if (faucetRes) return faucetRes
+    // if we reached here, either we used the faucet or user should call faucet on its own.
+    let txHash = ''
+    // simulate tx to detect revert
+    const canTopOrError = await this.proxyContract.methods
+      .topWallet(address)
+      .call()
+      .then(_ => true)
+      .catch(e => e)
 
-    let txHash
+    if (canTopOrError !== true) {
+      logger.debug('Topwallet will revert, skipping', { address, canTopOrError })
+      throw new Error('Topwallet will revert, probably user passed limit')
+    }
+
     try {
       const onTransactionHash = hash => {
         logger.debug('Topwallet got txhash:', { hash, address })
@@ -500,11 +546,13 @@ export class Wallet {
       const txPromise = this.sendTransaction(
         this.proxyContract.methods.topWallet(address),
         { onTransactionHash },
-        {},
+        { gas: 500000 },
         true,
         logger
       )
+
       let res = await txPromise
+
       logger.debug('Topwallet result:', { txHash, address, res })
       return res
     } catch (e) {
@@ -513,11 +561,31 @@ export class Wallet {
     }
   }
 
-  async topWalletFaucet(address, logger = log) {
+  async topWalletFaucet(address, customLogger = null) {
+    const logger = customLogger || this.log
+
     try {
       const canTop = await this.faucetContract.methods.canTop(address).call()
+
+      logger.debug('topWalletFaucet canTop result:', { address, canTop })
+
       if (canTop === false) {
         return false
+      }
+
+      let userBalance = web3Utils.toBN(await this.web3.eth.getBalance(address))
+      let faucetTxCost = web3Utils.toBN('150000').mul(web3Utils.toBN(this.gasPrice))
+
+      logger.debug('topWalletFaucet:', {
+        address,
+        userBalance: userBalance.toString(),
+        faucetTxCost: faucetTxCost.toString()
+      })
+
+      // user can't call faucet directly
+      if (userBalance.gte(faucetTxCost)) {
+        logger.debug('User has enough gas to call faucet', { address })
+        return true
       }
 
       let encodedCall = this.web3.eth.abi.encodeFunctionCall(
@@ -533,11 +601,12 @@ export class Wallet {
         },
         [address]
       )
-      const transaction = await this.proxyContract.methods.genericCall(this.faucetContract._address, encodedCall, 0)
 
-      const txPromise = this.sendTransaction(transaction, {}, { gas: 200000 }, true, logger)
-
+      const transaction = this.proxyContract.methods.genericCall(this.faucetContract._address, encodedCall, 0)
+      const onTransactionHash = hash => void logger.debug('topWalletFaucet got txhash:', { hash, address })
+      const txPromise = this.sendTransaction(transaction, { onTransactionHash }, { gas: 500000 }, true, logger)
       let res = await txPromise
+
       logger.debug('topWalletFaucet result:', { address, res })
       return res
     } catch (e) {
@@ -545,7 +614,10 @@ export class Wallet {
       throw e
     }
   }
-  async fishMulti(toFish: Array<string>, logger = log): Promise<TransactionReceipt> {
+
+  async fishMulti(toFish: Array<string>, customLogger = null): Promise<TransactionReceipt> {
+    const logger = customLogger || this.log
+
     try {
       let encodedCall = this.web3.eth.abi.encodeFunctionCall(
         {
@@ -560,9 +632,12 @@ export class Wallet {
         },
         [toFish]
       )
+
       logger.info('fishMulti sending tx', { encodedCall, toFish, ubischeme: this.UBIContract._address })
+
       const transaction = await this.proxyContract.methods.genericCall(this.UBIContract._address, encodedCall, 0)
       const tx = await this.sendTransaction(transaction, {}, { gas: 2000000 }, false, logger)
+
       logger.info('fishMulti success', { toFish, tx: tx.transactionHash })
       return tx
     } catch (exception) {
@@ -580,7 +655,9 @@ export class Wallet {
    * @param {*} logger
    * @returns
    */
-  async transferWalletGooDollars(to, value, logger = log): Promise<TransactionReceipt> {
+  async transferWalletGooDollars(to, value, customLogger = null): Promise<TransactionReceipt> {
+    const logger = customLogger || this.log
+
     try {
       let encodedCall = this.web3.eth.abi.encodeFunctionCall(
         {
@@ -600,8 +677,10 @@ export class Wallet {
         [to, value]
       )
       logger.info('transferWalletGooDollars sending tx', { encodedCall, to, value })
+
       const transaction = await this.proxyContract.methods.genericCall(this.tokenContract._address, encodedCall, 0)
       const tx = await this.sendTransaction(transaction, {}, { gas: 200000 }, false, logger)
+
       logger.info('transferWalletGooDollars success', { to, value, tx: tx.transactionHash })
       return tx
     } catch (exception) {
@@ -621,6 +700,8 @@ export class Wallet {
    * @returns {Promise<number>}
    */
   async getBalance(): Promise<number> {
+    const { log } = this
+
     return this.getAddressBalance(this.address)
       .then(b => parseFloat(web3Utils.fromWei(b)))
       .catch(e => {
@@ -647,53 +728,77 @@ export class Wallet {
     txCallbacks: PromiEvents = {},
     { gas, gasPrice }: GasValues = { gas: undefined, gasPrice: undefined },
     retry = true,
-    logger = log
+    customLogger = null
   ) {
     let currentAddress, txHash
     const uuid = Crypto.randomBytes(5).toString('base64')
+    const logger = customLogger || this.log
+
     try {
       const { onTransactionHash, onReceipt, onConfirmation, onError } = txCallbacks
+
       gas =
         gas ||
         (await tx
           .estimateGas()
-          .then(gas => gas + 200000) //buffer for proxy contract, reimburseGas?
+          .then(gas => parseInt(gas) + 200000) //buffer for proxy contract, reimburseGas?
           .catch(e => {
             logger.warn('Failed to estimate gas for tx', e.message, e)
             return defaultGas
           }))
 
       // adminwallet contract might give wrong gas estimates, so if its more than block gas limit reduce it to default
-      if (gas > 8000000) gas = defaultGas
-      gasPrice = gasPrice || defaultGasPrice
+      if (gas > 8000000) {
+        gas = defaultGas
+      }
+
+      gasPrice = gasPrice || this.gasPrice
 
       logger.debug('getting tx lock:', { uuid })
+
       const { nonce, release, fail, address } = await this.txManager.lock(this.filledAddresses)
+
       logger.debug('got tx lock:', { uuid, address })
 
       let balance = NaN
-      if (conf.env === 'development') {
+
+      if (this.conf.env === 'development') {
         balance = await this.web3.eth.getBalance(address)
       }
+
       currentAddress = address
-      logger.debug(`sending tx from: ${address} | nonce: ${nonce}`, { uuid, balance, gas, gasPrice })
+
+      logger.debug(`sending tx from:`, { address, nonce, uuid, balance, gas, gasPrice })
+
       let txPromise = new Promise((res, rej) => {
         tx.send({ gas, gasPrice, chainId: this.networkId, nonce, from: address })
           .on('transactionHash', h => {
             release()
+
             txHash = h
             logger.debug('got tx hash:', { uuid, txHash })
-            onTransactionHash && onTransactionHash(h)
+
+            if (onTransactionHash) {
+              onTransactionHash(h)
+            }
+          })
+          .on('sent', payload => {
+            logger.debug('tx sent:', { txHash, payload })
           })
           .on('receipt', r => {
             logger.debug('got tx receipt:', { uuid })
-            onReceipt && onReceipt(r)
+
+            if (onReceipt) {
+              onReceipt(r)
+            }
+
             res(r)
           })
           .on('confirmation', c => onConfirmation && onConfirmation(c))
           .on('error', async e => {
             if (isFundsError(e)) {
               balance = await this.web3.eth.getBalance(address)
+
               logger.warn('sendTransaciton funds issue retry', {
                 errMessage: e.message,
                 nonce,
@@ -702,8 +807,10 @@ export class Wallet {
                 address,
                 balance
               })
+
               sendSlackAlert({ msg: 'admin account funds low', address, balance })
               await this.txManager.unlock(address)
+
               try {
                 res(await this.sendTransaction(tx, txCallbacks, { gas, gasPrice }, retry, logger))
               } catch (e) {
@@ -712,6 +819,7 @@ export class Wallet {
               }
             } else if (isNonceError(e)) {
               let netNonce = parseInt(await this.web3.eth.getTransactionCount(address))
+
               logger.warn('sendTransaciton nonce failure retry', {
                 errMessage: e.message,
                 nonce,
@@ -720,9 +828,11 @@ export class Wallet {
                 address,
                 newNonce: netNonce
               })
+
               await this.txManager.unlock(address, netNonce)
+
               try {
-                res(await this.sendTransaction(tx, txCallbacks, { gas, gasPrice }, retry, logger))
+                await this.sendTransaction(tx, txCallbacks, { gas, gasPrice }, retry, logger).then(res)
               } catch (e) {
                 await this.txManager.unlock(address)
                 rej(e)
@@ -735,16 +845,18 @@ export class Wallet {
           })
       })
 
-      const res = await Promise.race([txPromise, requestTimeout(FUSE_TX_TIMEOUT, 'fuse tx timeout')])
-      return res
+      const response = await withTimeout(txPromise, FUSE_TX_TIMEOUT, 'fuse tx timeout')
+
+      return response
     } catch (e) {
       await this.txManager.unlock(currentAddress)
+
       if (retry && e.message.includes('fuse tx timeout')) {
         logger.warn('sendTransaction timeout retrying:', { uuid, txHash })
         return this.sendTransaction(tx, txCallbacks, { gas, gasPrice }, false, logger)
       }
-      logger.error('sendTransaction error:', e.message, e, { from: currentAddress, uuid })
 
+      logger.error('sendTransaction error:', e.message, e, { from: currentAddress, uuid })
       throw new Error(e)
     }
   }
@@ -768,12 +880,16 @@ export class Wallet {
     { gas, gasPrice }: GasValues = { gas: undefined, gasPrice: undefined }
   ) {
     let currentAddress
+    const { log } = this
+
     try {
       const { onTransactionHash, onReceipt, onConfirmation, onError } = txCallbacks
+
       gas = gas || defaultGas
-      gasPrice = gasPrice || defaultGasPrice
+      gasPrice = gasPrice || this.gasPrice
 
       const { nonce, release, fail, address } = await this.txManager.lock(this.filledAddresses)
+
       log.debug('sendNative', { nonce, gas, gasPrice })
       currentAddress = address
 
@@ -781,21 +897,30 @@ export class Wallet {
         this.web3.eth
           .sendTransaction({ gas, gasPrice, chainId: this.networkId, nonce, ...params, from: address })
           .on('transactionHash', h => {
-            onTransactionHash && onTransactionHash(h)
+            if (onTransactionHash) {
+              onTransactionHash(h)
+            }
+
             release()
           })
           .on('receipt', r => {
-            onReceipt && onReceipt(r)
+            if (onReceipt) {
+              onReceipt(r)
+            }
+
             res(r)
           })
           .on('confirmation', c => {
-            onConfirmation && onConfirmation(c)
+            if (onConfirmation) {
+              onConfirmation(c)
+            }
           })
           .on('error', async e => {
             const { message } = e
 
             if (isNonceError(e)) {
               let netNonce = parseInt(await this.web3.eth.getTransactionCount(address))
+
               log.warn('sendNative nonce failure retry', message, e, {
                 params,
                 nonce,
@@ -804,16 +929,22 @@ export class Wallet {
                 address,
                 newNonce: netNonce
               })
+
               await this.txManager.unlock(address, netNonce)
+
               try {
-                res(await this.sendNative(params, txCallbacks, { gas, gasPrice }))
+                await this.sendNative(params, txCallbacks, { gas, gasPrice }).then(res)
               } catch (e) {
                 await this.txManager.unlock(address)
                 rej(e)
               }
             } else {
               fail()
-              onError && onError(e)
+
+              if (onError) {
+                onError(e)
+              }
+
               log.error('sendNative failed', message, e)
               rej(e)
             }
@@ -824,135 +955,4 @@ export class Wallet {
       throw new Error(e)
     }
   }
-
-  /**
-   * Helper function to handle a tx Send call
-   * @param tx
-   * @param {object} promiEvents
-   * @param {function} promiEvents.onTransactionHash
-   * @param {function} promiEvents.onReceipt
-   * @param {function} promiEvents.onConfirmation
-   * @param {function} promiEvents.onError
-   * @param {object} gasValues
-   * @param {number} gasValues.gas
-   * @param {number} gasValues.gasPrice
-   * @returns {Promise<Promise|Q.Promise<any>|Promise<*>|Promise<*>|Promise<*>|*>}
-   */
-  async sendTransactionMainnet(
-    tx: any,
-    txCallbacks: PromiEvents = {},
-    { gas, gasPrice }: GasValues = { gas: undefined, gasPrice: undefined },
-    forceAddress: string
-  ) {
-    let currentAddress
-    try {
-      const { onTransactionHash, onReceipt, onConfirmation, onError } = txCallbacks
-      gas =
-        gas ||
-        (await tx
-          .estimateGas()
-          .then(gas => gas + 200000) //buffer for proxy contract, reimburseGas?, and low gas unexpected failures
-          .catch(e => {
-            log.warn('Failed to estimate gas for tx mainnet', e.message, e)
-            return defaultGas
-          }))
-
-      //adminwallet contract might give wrong gas estimates, so if its more than block gas limit reduce it to default
-      if (gas > 8000000) gas = defaultGas
-      // gasPrice = gasPrice || Math.min(await this.mainnetWeb3.eth.getGasPrice(), this.maxMainnetGasPrice)
-      gasPrice = gasPrice || defaultRopstenGasPrice
-
-      const uuid = Crypto.randomBytes(5).toString('base64')
-      log.debug('getting tx lock mainnet:', { uuid })
-      const { nonce, release, fail, address } = await this.mainnetTxManager.lock(forceAddress || this.mainnetAddresses)
-      log.debug('got tx lock mainnet:', { uuid, address })
-
-      let balance = NaN
-      if (conf.env === 'development') {
-        balance = await this.mainnetWeb3.eth.getBalance(address)
-      }
-      currentAddress = address
-      log.debug(`sending  tx mainnet from: ${address} | nonce: ${nonce}`, {
-        network: this.networkIdMainNet,
-        uuid,
-        balance,
-        gas,
-        gasPrice
-      })
-      return new Promise((res, rej) => {
-        tx.send({
-          // type: '0x2',
-          gas,
-          gasPrice,
-          // maxFeePerGas: gasPrice,
-          // maxPriorityFeePerGas: web3Utils.toWei('1', 'gwei'),
-          // chainId: this.networkIdMainNet,
-          nonce,
-          from: address
-        })
-          .on('transactionHash', h => {
-            release()
-            log.debug('got tx hash mainnet:', { uuid })
-            onTransactionHash && onTransactionHash(h)
-          })
-          .on('receipt', r => {
-            log.debug('got tx receipt mainnet:', { uuid })
-            onReceipt && onReceipt(r)
-            res(r)
-          })
-          .on('confirmation', c => onConfirmation && onConfirmation(c))
-          .on('error', async exception => {
-            const { message } = exception
-
-            if (isFundsError(exception)) {
-              balance = await this.mainnetWeb3.eth.getBalance(address)
-              log.warn('sendTransaciton funds issue retry mainnet', {
-                errMessage: message,
-                nonce,
-                gas,
-                gasPrice,
-                address,
-                balance
-              })
-              sendSlackAlert({ msg: 'admin account funds low mainnet', address, balance })
-              await this.mainnetTxManager.unlock(address)
-              try {
-                res(await this.sendTransaction(tx, txCallbacks, { gas, gasPrice }))
-              } catch (e) {
-                await this.mainnetTxManager.unlock(address)
-                rej(e)
-              }
-            } else if (isNonceError(exception)) {
-              let netNonce = parseInt(await this.mainnetWeb3.eth.getTransactionCount(address))
-              log.warn('sendTransaciton nonce failure retry mainnet', {
-                errMessage: message,
-                nonce,
-                gas,
-                gasPrice,
-                address,
-                newNonce: netNonce
-              })
-              await this.mainnetTxManager.unlock(address, netNonce)
-              try {
-                res(await this.sendTransactionMainnet(tx, txCallbacks, { gas, gasPrice }))
-              } catch (e) {
-                await this.mainnetTxManager.unlock(address)
-                rej(e)
-              }
-            } else {
-              fail()
-              onError && onError(exception)
-              log.error('sendTransaction error mainnet:', message, exception, { from: address, uuid })
-              rej(exception)
-            }
-          })
-      })
-    } catch (e) {
-      await this.mainnetTxManager.unlock(currentAddress)
-      throw new Error(e)
-    }
-  }
 }
-
-const AdminWallet = new Wallet(conf.mnemonic)
-export default AdminWallet
